requirements.txt
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
astroid==2.2.5
attrs==19.3.0
autoit==0.1
autopep8==1.4.4
backcall==0.2.0
bleach==3.1.5
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
comtypes==1.1.7
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1596868620883/work
ipython-genutils==0.2.0
isort==4.3.21
jedi @ file:///C:/ci/jedi_1596472767194/work
Jinja2==2.11.2
json5==0.9.5
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
keyboard==0.13.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib==3.3.1
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mouse==0.7.1
MouseInfo==0.1.3
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
notebook @ file:///C:/ci/notebook_1596837179121/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233945180/work
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
PyAutoGUI==0.9.50
PyAutoIt==0.6.0
pycodestyle==2.5.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyDirectInput==1.0.2
pydivert==2.1.0
PyGetWindow==0.0.8
Pygments==2.6.1
pylint==2.3.1
PyMsgBox==1.0.8
pynput==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
PyTweening==1.0.3
pywin32==227
pywinauto==0.6.8
pyWinhook==1.6.2
pywinpty==0.5.7
pyzmq==19.0.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlib==0.21
zipp==3.1.0
